Added pipeline phase to fetch tags using JdbcIO and phases to merge resource and tags collection

+++ pipelines/batch/src/main/java/org/openmrs/analytics/FhirEtlOptions.java
@@ -103,7 +103,7 @@ public interface FhirEtlOptions extends PipelineOptions {
   void setJdbcDriverClass(String value);
 
   @Description("JDBC maximum pool size")
-  @Default.Integer(50)
+  @Default.Integer(20)
   int getJdbcMaxPoolSize();
 
   void setJdbcMaxPoolSize(int value);
